requirements.txt
#------------------------------------------
# Requirements with Version Specifiers
#------------------------------------------
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
PyJWT==1.0.0